1. correct HardwareMap instance passed to subsystems 2. removed bloat caused by example files 3. split auton and teleop opmodes into seperate folders 4. created a module for MeepMeep testing
