+++ MeepMeepTesting/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+    id 'java-library'
+}
+java {
+    sourceCompatibility JavaVersion.VERSION_11
+    targetCompatibility JavaVersion.VERSION_11
+}
