+++ settings.gradle
@@ -1,2 +1,4 @@
 include ':FtcRobotController'
 include ':TeamCode'
+include ':TeamCode:meepmeepingtesting'
+include ':MeepMeepTesting'
